chore: Update Cart component to include Checkout component for finalizing the order
